Universal framework

+++ libs/SalesforceSDKCore/SalesforceSDKCore/Classes/Test/SFSDKTestRequestListener.m
@@ -121,7 +121,7 @@ NSString* const kTestRequestStatusDidTimeout = @"didTimeout";
     } else if (_serviceType == SFAccountManagerServiceTypeOAuth) {
         _authMgr.coordinator.delegate = self;
     } else {
-        NSAssert1(NO, @"Service type '%d' is not supported as a service object.", _serviceType);
+        NSAssert1(NO, @"Service type '%lu' is not supported as a service object.", (unsigned long)_serviceType);
     }
 }
